@@ -12,36 +12,36 @@
     "postbuild": "next-sitemap --config next-sitemap.config.js"
   },
   "dependencies": {
-    "@headlessui/react": "^2.1.1",
-    "@next/third-parties": "^14.2.4",
-    "@sendgrid/mail": "^8.1.3",
-    "@tailwindcss/aspect-ratio": "^0.4.0",
-    "@tanstack/react-query": "^5.62.1",
-    "dompurify": "^3.1.5",
-    "dotenv": "^16.0.0",
-    "html-react-parser": "^5.1.10",
-    "next": "^14.0.3",
-    "next-seo": "^6.5.0",
+    "@headlessui/react": "^2.2.7",
+    "@next/third-parties": "^14.2.31",
+    "@sendgrid/mail": "^8.1.5",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tanstack/react-query": "^5.84.1",
+    "dompurify": "^3.2.6",
+    "dotenv": "^16.6.1",
+    "html-react-parser": "^5.2.6",
+    "next": "^14.2.31",
+    "next-seo": "^6.8.0",
     "next-sitemap": "^4.2.3",
-    "node-fetch": "^3.2.0",
-    "preact": "^10.6.4",
+    "node-fetch": "^3.3.2",
+    "preact": "^10.27.0",
     "prop-types": "^15.8.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-ga": "^3.3.0",
-    "sass": "^1.69.5",
-    "sharp": "^0.33.4"
+    "react-ga": "^3.3.1",
+    "sass": "^1.89.2",
+    "sharp": "^0.33.5"
   },
   "devDependencies": {
     "@svgr/cli": "^8.1.0",
     "@types/gtag.js": "^0.0.20",
-    "@types/node": "^20.14.9",
-    "@types/react": "^18.2.0",
-    "autoprefixer": "^10.4.2",
+    "@types/node": "^20.19.9",
+    "@types/react": "^18.3.23",
+    "autoprefixer": "^10.4.21",
     "eslint": "8.0.0",
     "eslint-config-next": "14.2.4",
-    "postcss": "^8.4.5",
-    "tailwindcss": "^3.0.15"
+    "postcss": "^8.5.6",
+    "tailwindcss": "^3.4.17"
   },
   "packageManager": "pnpm@10.11.0+sha512.6540583f41cc5f628eb3d9773ecee802f4f9ef9923cc45b69890fb47991d4b092964694ec3a4f738a420c918a333062c8b925d312f42e4f0c263eb603551f977"
 }
